Refactored with more checks and using gweis for store and calculation

+++ contracts/structs/Fund.sol
@@ -3,5 +3,5 @@ pragma solidity >=0.6.11 <0.7.0;
 
 struct Fund {
     address from;
-    uint amount;
+    uint64 amount;
 }
